requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.44
botocore==1.35.44
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
django-proxy==1.3.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==28.4.1
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
jwcrypto==1.5.6
multidict==6.1.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
twilio==9.3.2
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
whitenoise==6.7.0
yarl==1.13.1
django-revproxy==0.12.0